+++ src/core/services/api.ts
@@ -1,5 +1,6 @@
 import * as Rx from 'rx';
 import * as angular from 'angular';
+import * as _ from 'lodash';
 import 'ng-file-upload';
 
 import {GenError} from '../errors/error';
@@ -15,10 +16,12 @@ const angularModule: angular.IModule = angular.module('resolwe.services.api', [
 
 export enum UploadEventType {
     PROGRESS = 'progress',
+    RETRYING = 'retrying',
     RESULT = 'result',
 }
 
 export type UploadEvent<T> = { progress: ProgressEvent, type: UploadEventType.PROGRESS } |
+                             { type: UploadEventType.RETRYING } |
                              { result: T, type: UploadEventType.RESULT };
 
 /**
@@ -27,6 +30,13 @@ export type UploadEvent<T> = { progress: ProgressEvent, type: UploadEventType.PR
  * actual API class.
  */
 export class APIServiceBase {
+    // Note that this is configurable in provider
+    public consecutiveAutoretryAttempts = 5;
+    public CHUNK_SIZE = 1 * 1024 * 1024; // 1MB
+    public RETRY_DELAY_INCREMENT = 500;
+    public MAX_RETRY_DELAY = 5000;
+    public ERROR_GROUPING_TIME = () => this.RETRY_DELAY_INCREMENT * this.consecutiveAutoretryAttempts * 2;
+
     // Note that this connection property is not initialized anywhere as it will
     // be initialized by the actual API which is mixed in by the provider.
     public connection: Connection;
@@ -37,10 +47,15 @@ export class APIServiceBase {
 
     constructor(Upload: angular.angularFileUpload.IUploadService,
                 $q: angular.IQService,
-                $http: angular.IHttpService) {
+                $http: angular.IHttpService,
+                consecutiveAutoretryAttempts?: number) {
         this._upload = Upload;
         this._q = $q;
         this._http = $http;
+
+        if (!_.isUndefined(consecutiveAutoretryAttempts)) {
+            this.consecutiveAutoretryAttempts = consecutiveAutoretryAttempts;
+        }
     }
 
     /**
@@ -66,6 +81,8 @@ export class APIServiceBase {
         };
 
         return Rx.Observable.create<UploadEvent<T>>((observer) => {
+            observer.onNext({ type: UploadEventType.RETRYING }); // Note: First one of these is skipped below.
+
             const rejectableResumeSizePromise = this._q.defer<number>();
             const fileUpload = this._upload.upload<T>({
                 url: url,
@@ -86,7 +103,7 @@ export class APIServiceBase {
                     rejectableResumeSizePromise.resolve(resumeSizePromise);
                     return rejectableResumeSizePromise.promise;
                 },
-                resumeChunkSize: '1MB',
+                resumeChunkSize: this.CHUNK_SIZE,
                 data: data,
             });
 
@@ -103,6 +120,38 @@ export class APIServiceBase {
                 fileUpload.abort();
                 rejectableResumeSizePromise.reject();
             };
+        })
+        .retryWhen((errors) => {
+            return errors
+                .filter((error) => error && error.xhrStatus !== 'abort')
+                .timeInterval()
+                .scan((accumulated, value) => {
+                    const error = value.value;
+                    const timeSincePrevious = value.interval;
+
+                    let consecutiveErrors = accumulated.consecutiveErrors + 1;
+                    if (timeSincePrevious > this.ERROR_GROUPING_TIME()) consecutiveErrors = 1;
+
+                    const retry = consecutiveErrors <= this.consecutiveAutoretryAttempts;
+                    const delay = Math.min(consecutiveErrors * this.RETRY_DELAY_INCREMENT, this.MAX_RETRY_DELAY);
+
+                    return { error, consecutiveErrors, timeSincePrevious, retry, delay };
+                }, { error: null, consecutiveErrors: 0, timeSincePrevious: 0, retry: false, delay: 0 })
+                .flatMap(({retry, delay, error}) => {
+                    if (!retry) { // Stop retrying after a while and return unwrapped error
+                        return Rx.Observable.throw(error);
+                    }
+                    return Rx.Observable.just(error).delay(delay);
+                })
+                .do((error) => {
+                    console.log("Retrying upload after an error", error);
+                });
+        })
+        .skip(1) // Skip initial 'retrying' event.
+        .filter((event) => {
+            // If a retry request fails, it would remove the progress bar until it
+            // succeeds again. With this filter we can keep the progress bar anyway.
+            return !(event.type === UploadEventType.PROGRESS && event.progress.loaded === 0 && event.progress.total === 0);
         });
     }
 
@@ -143,6 +192,9 @@ export class APIServiceBase {
  *         REST_URL,
  *         WEBSOCKET_URL
  *     );
+ *
+ *     // Configure upload auto-retries to infinity
+ *     apiProvider.setConsecutiveAutoretryAttempts(Infinity);
  * });
  * ```
  */
@@ -152,6 +204,7 @@ export class APIServiceProvider {
     private _connection: Connection;
     private _restUri: string;
     private _websocketUri: string;
+    private _consecutiveAutoretryAttempts: number;
 
     public setAPI(api: typeof ResolweApi,
                   connection: Connection,
@@ -163,11 +216,15 @@ export class APIServiceProvider {
         this._websocketUri = websocketUri;
     }
 
+    public setConsecutiveAutoretryAttempts(retries: number) {
+        this._consecutiveAutoretryAttempts = retries;
+    }
+
     // @ngInject
     public $get(Upload: angular.angularFileUpload.IUploadService,
                 $q: angular.IQService,
                 $http: angular.IHttpService) {
-        // TODO: Use error notifciation service instead.
+        // TODO: Use error notification service instead.
         if (!this._api) throw new GenError("API not configured.");
 
         // Mix together the API and the APIServiceBase.
@@ -176,7 +233,7 @@ export class APIServiceProvider {
             // Arguments for the API part.
             [this._connection, this._restUri, this._websocketUri],
             // Arguments for APIServiceBase part.
-            [Upload, $q, $http]
+            [Upload, $q, $http, this._consecutiveAutoretryAttempts]
         );
     }
 }